requirements.txt
pycrypto
moviepy
sqlalchemy
chardet
flask==0.12.2
tornado
Pillow